🎨 BadgeRepository 잘못된 id 타입 수정

+++ adevspoon-domain/src/main/kotlin/com/adevspoon/domain/member/repository/BadgeRepository.kt
@@ -3,5 +3,5 @@ package com.adevspoon.domain.member.repository
 import com.adevspoon.domain.member.domain.BadgeEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface BadgeRepository: JpaRepository<BadgeEntity, Long> {
+interface BadgeRepository: JpaRepository<BadgeEntity, Int> {
 }
